% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2-latlng.R
\name{as_s2_latlng}
\alias{as_s2_latlng}
\alias{as_s2_latlng.s2_latlng}
\alias{as_s2_latlng.s2_point}
\alias{as_s2_latlng.numeric}
\alias{as_s2_latlng.matrix}
\alias{as.data.frame.s2_latlng}
\alias{as.matrix.s2_latlng}
\title{Create an s2_latlng vector}
\usage{
as_s2_latlng(lat, ...)

\method{as_s2_latlng}{s2_latlng}(lat, ...)

\method{as_s2_latlng}{s2_point}(lat, ...)

\method{as_s2_latlng}{numeric}(lat, lng, ...)

\method{as_s2_latlng}{matrix}(lat, ...)

\method{as.data.frame}{s2_latlng}(x, ...)

\method{as.matrix}{s2_latlng}(x, ...)
}
\arguments{
\item{lat, lng}{Vectors of latitude and longitude values in degrees.}

\item{...}{Unused}

\item{x}{A \code{\link[=as_s2_latlng]{as_s2_latlng()}} vector}
}
\value{
A \code{\link[=new_s2_xptr]{new_s2_xptr()}} with class s2_latlng
}
\description{
This class represents a latitude and longitude on the Earth's surface.
Most calculations in S2 convert this to a \code{\link[=as_s2_point]{as_s2_point()}}, which is a
unit vector representation of this value.
}
\examples{
as_s2_latlng(45, -64) # Halifax, Nova Scotia!
as.data.frame(as_s2_latlng(45, -64))

}
