% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkb.R
\name{as_wkb}
\alias{as_wkb}
\alias{as_wkb.s2latlng}
\alias{as_wkb.s2polyline}
\alias{as_wkb.s2polygon}
\alias{s2latlng.wk_wkb}
\alias{s2polyline.wk_wkb}
\alias{s2polygon.wk_wkb}
\title{Convert to and from well-known binary}
\usage{
as_wkb(x, ...)

\method{as_wkb}{s2latlng}(x, ...)

\method{as_wkb}{s2polyline}(x, ...)

\method{as_wkb}{s2polygon}(x, ...)

\method{s2latlng}{wk_wkb}(lat, ...)

\method{s2polyline}{wk_wkb}(x, ...)

\method{s2polygon}{wk_wkb}(x, ...)
}
\arguments{
\item{x, lng}{Objects}

\item{...}{Unused}
}
\value{
A well-known binary vector (\code{list()} of \code{raw()})
}
\description{
Convert to and from well-known binary
}
