% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cell_center}
\alias{cell_center}
\title{cell_center}
\usage{
cell_center(lat, lng)
}
\arguments{
\item{lat}{latitude of interest (between -90 and +90)}

\item{lng}{longitude of interest (between -180 and +180)}
}
\description{
Get center of S2 cell containing a specified point
}
\details{
\code{cell_center} takes a single (lat,lng) point on the sphere, finds the
S2 leaf cell that contains this point and returns the center of this cell.
In S2 leaf cells have a very fine resolution, so the output is expected to
be very close to the input, and differences are typically on something like
the 5th decimal place.
}
\examples{
cell_center(57.0139595, 9.988967)

}

