\name{interval_test}
\alias{interval_test}
\title{Interval test}
\description{Test whether a number is inside a given interval (dummy function to test package).}
\usage{
interval_test(x, lower = -pi, upper = pi)
}
\arguments{
\item{x}{Numeric. Point to be tested.}

\item{lower}{Numeric. Lower limit of interval.}

\item{upper}{Numeric. Upper limit of interval.}
}
\value{
Logical.
}
