% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs2-package.R
\docType{package}
\name{libs2-package}
\alias{libs2}
\alias{libs2-package}
\title{libs2: Spherical Geometry Operators, Using the s2geometry.io Library}
\description{
Geometry operations on the sphere using s2geometry.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-spatial/libs2/ https://s2geometry.io/}
  \item Report bugs at \url{https://github.com/r-spatial/libs2/issues}
}

}
\author{
\strong{Maintainer}: Edzer Pebesma \email{edzer.pebesma@uni-muenster.de} (\href{https://orcid.org/0000-0001-8049-7069}{ORCID})

Authors:
\itemize{
  \item Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})
  \item Ege Rubak \email{rubak@math.aau.dk}
}

Other contributors:
\itemize{
  \item Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu} (configure and cleanup script) [contributor]
  \item  Google, Inc. (Original s2geometry.io source code) [copyright holder]
}

}
\keyword{internal}
