% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc_raw}
\alias{nc_raw}
\title{Geometry of the nc dataset of package sf}
\format{
A list with sfg objects
}
\source{
\url{http://r-spatial.github.io/sf/}
}
\usage{
nc_raw
}
\description{
The raw output of st_geometry(nc) when read with package sf
}
\examples{
# Created with:
# library(sf)
# read_sf(system.file("gpkg/nc.gpkg", package="sf"), check_ring_dir = FALSE) \%>\%
#	st_geometry() -> nc_raw
}
\keyword{datasets}
