% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2-xptr.R
\name{new_s2_xptr}
\alias{new_s2_xptr}
\alias{validate_s2_xptr}
\alias{rep.s2_xptr}
\alias{rep_len.s2_xptr}
\title{Create vectors of XPtr objects}
\usage{
new_s2_xptr(x = list(), class = character())

validate_s2_xptr(x)

\method{rep}{s2_xptr}(x, ...)

rep_len.s2_xptr(x, ...)
}
\arguments{
\item{x}{A bare \code{list()} of external pointers}

\item{class}{A character vector subclass}

\item{...}{Unused}
}
\value{
An object of class s2_xptr
}
\description{
Create vectors of XPtr objects
}
