% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2xptr.R
\name{new_s2xptr}
\alias{new_s2xptr}
\alias{validate_s2xptr}
\title{Create S2 vectors of XPtr objects}
\usage{
new_s2xptr(x = list(), class = character())

validate_s2xptr(x)
}
\arguments{
\item{x}{A bare \code{list()} of external pointers}

\item{class}{A character vector subclass}
}
\value{
An object of class s2xptr
}
\description{
Create S2 vectors of XPtr objects
}
