% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2.R
\docType{package}
\name{s2-package}
\alias{s2}
\alias{s2-package}
\title{s2: Google's S2 Library for Geometry on the Sphere}
\description{
R bindings for Google's s2 library for geometric calculations on
    the sphere.
}
\details{
This package aims at providing bindings for the C++ library s2 for geometry
on the sphere. The C++ library was originally developed by Google under the
Apache license. At the moment very few things are exposed and the main
concern is to make the C++ code compile with the R toolchain on the three
major platforms and have the package accepted on CRAN. Once this proof of
concept has been achieved more of the underlying features will be made
available to the useR.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/spatstat/s2}
  \item \url{https://code.google.com/archive/p/s2-geometry-library/}
}

}
\author{
\strong{Maintainer}: Ege Rubak \email{rubak@math.aau.dk}

Authors:
\itemize{
  \item Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu} (configure and cleanup script)
  \item Edzer Pebesma \email{edzer.pebesma@uni-muenster.de} (Interface to sf)
}

Other contributors:
\itemize{
  \item  Google, Inc. (Original s2 source code) [copyright holder]
}

}
