% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2-transformers.R
\name{s2_boundary}
\alias{s2_boundary}
\alias{s2_centroid}
\alias{s2_closest_point}
\alias{s2_minimum_clearance_line_between}
\alias{s2_difference}
\alias{s2_sym_difference}
\alias{s2_intersection}
\alias{s2_union}
\alias{s2_snap_to_grid}
\alias{s2_union_agg}
\alias{s2_centroid_agg}
\title{S2 Geography Transformations}
\usage{
s2_boundary(x)

s2_centroid(x)

s2_closest_point(x, y)

s2_minimum_clearance_line_between(x, y)

s2_difference(x, y, model = s2_model_default(), snap_level = s2_snap_default())

s2_sym_difference(
  x,
  y,
  model = s2_model_default(),
  snap_level = s2_snap_default()
)

s2_intersection(
  x,
  y,
  model = s2_model_default(),
  snap_level = s2_snap_default()
)

s2_union(
  x,
  y = NULL,
  model = s2_model_default(),
  snap_level = s2_snap_default()
)

s2_snap_to_grid(x)

s2_union_agg(
  x,
  model = s2_model_default(),
  snap_level = s2_snap_default(),
  na.rm = FALSE
)

s2_centroid_agg(x, na.rm = FALSE)
}
\arguments{
\item{x}{An object that can be interpreted as an S2 type.}

\item{y}{An object that can be interpreted as an S2 type.}

\item{model}{integer; specifies boundary model. See \code{\link[=s2_model_default]{s2_model_default()}}.}

\item{snap_level}{integer; if positive, specifies the snap level. See \code{\link[=s2_snap_default]{s2_snap_default()}}.}

\item{na.rm}{For aggregate calculations use \code{na.rm = TRUE}
to drop missing values.}
}
\description{
S2 Geography Transformations
}
\section{Model}{

The geometry model indicates whether a geometry includes its boundaries.
Boundaries of line geometries are its end points.
OPEN geometries do not contain their boundary (model = 0); CLOSED
geometries (model = 2) contain their boundary; HALF-CLOSED geometries
contain, like, half of their boundaries, such that when two polygons
do not overlap or two lines do not cross, no point exist that belong to
more than one of the geometries. (This latter form, half-closed, is
not present n the OpenGIS "simple feature access" (SFA) standard, or DE9-IM on
which that is based). A value of -1 does not set the model, leaving the
S2 default (HALF-CLOSED). The default values for \code{\link[=s2_contains]{s2_contains()}} (0)
and covers/coveredby (2) correspond to the SFA standard specification
of these operators.
}

\seealso{
BigQuery's geography function reference:
\itemize{
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_boundary}{ST_BOUNDARY}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_centroid}{ST_CENTROID}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_closestpoint}{ST_CLOSESTPOINT}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_difference}{ST_DIFFERENCE}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_intersection}{ST_INTERSECTION}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_union}{ST_UNION}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_snaptogrid}{ST_SNAPTOGRID}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_union_agg}{ST_UNION_AGG}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#s2_centroid_agg}{ST_CENTROID_AGG}
}
}
