% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2-cap.R
\name{s2_cap}
\alias{s2_cap}
\alias{s2_lat_lng_rect}
\title{compute S2Cap of each geography}
\usage{
s2_cap(x)

s2_lat_lng_rect(x)
}
\arguments{
\item{x}{object of class s2_geography}
}
\value{
\code{s2_cap} returns a bounding spherical cap that contains the region.
For points, polylines and geometry collections the bound may not be tight.

\code{s2_lat_lng_rect} returns a bounding latitude-longitude
rectangle that contains the region.  The bound may not be tight for points,
polylines and geometry collections. lng_lo values larger than lng_hi indicate
regions that span the antimeridian, see the Fiji example.
}
\description{
compute S2Cap of each geography
}
\examples{
s2_cap(s2_data_countries("Antarctica"))
s2_cap(s2_data_countries("Netherlands"))
s2_cap(s2_data_countries("Fiji"))
s2_lat_lng_rect(s2_data_countries("Antarctica"))
s2_lat_lng_rect(s2_data_countries("Netherlands"))
s2_lat_lng_rect(s2_data_countries("Fiji"))
}
