% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2-cell.R
\name{s2_cell}
\alias{s2_cell}
\alias{as_s2_cell}
\alias{as_s2_cell.character}
\alias{as_s2_cell.s2_geography}
\alias{as_s2_cell.s2_lnglat}
\alias{as_s2_cell.s2_point}
\title{Create S2 Cell vectors}
\usage{
s2_cell(x = character())

as_s2_cell(x, ...)

\method{as_s2_cell}{character}(x, ...)

\method{as_s2_cell}{s2_geography}(x, ...)

\method{as_s2_cell}{s2_lnglat}(x, ...)

\method{as_s2_cell}{s2_point}(x, ...)
}
\arguments{
\item{x}{The canonical S2 cell identifier as a character vector.}

\item{...}{Passed to methods}
}
\value{
An object of class s2_cell
}
\description{
Create S2 Cell vectors
}
\examples{
s2_cell("87754e64dffffff")

}
