% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2-cell.R
\name{s2_cell_is_valid}
\alias{s2_cell_is_valid}
\alias{s2_cell_to_lnglat}
\title{S2 cell operators}
\usage{
s2_cell_is_valid(x)

s2_cell_to_lnglat(x)
}
\arguments{
\item{x}{An \code{\link[=s2_cell]{s2_cell()}} vector}
}
\description{
S2 cell operators
}
