% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-predicates.R
\name{s2_contains}
\alias{s2_contains}
\alias{s2_coveredby}
\alias{s2_covers}
\alias{s2_disjoint}
\alias{s2_equals}
\alias{s2_intersects}
\alias{s2_intersectsbox}
\alias{s2_touches}
\alias{s2_within}
\alias{s2_dwithin}
\title{S2 Geography Predicates}
\usage{
s2_contains(x, y, model = 0)

s2_coveredby(x, y, model = 2)

s2_covers(x, y, model = 2)

s2_disjoint(x, y, ...)

s2_equals(x, y)

s2_intersects(x, y, model = -1)

s2_intersectsbox(x, lng1, lat1, lng2, lat2, detail = 1000)

s2_touches(x, y)

s2_within(x, y, ...)

s2_dwithin(x, y, distance, radius = s2earth_radius_meters())
}
\arguments{
\item{x}{An object that can be interpreted as an S2 type.}

\item{y}{An object that can be interpreted as an S2 type.}

\item{model}{integer or character; specify polygon and polyline model
as "OPEN" (or 0), "SEMI_OPEN" (or 1), or "CLOSED" (or 2)}

\item{...}{arguments passed on}

\item{lng1, lat1, lng2, lat2}{A latitude/longitude range}

\item{detail}{The number of points with which to approximate
non-geodesic edges.}

\item{distance}{A distance in meters on the surface of the earth}

\item{radius}{Radius of the earth. Defaults to the average radius of
the earth as defined by \code{\link[=s2earth_radius_meters]{s2earth_radius_meters()}}.}
}
\description{
S2 Geography Predicates
}
\details{
The geometry model indicates whether a geometry includes its boundaries.
Boundaries of line geometries are its end points.
OPEN geometries do not contain their boundary (model = 0); CLOSED
geometries (model = 2) contain their boundary; HALF-CLOSED geometries
contain, like, half of their boundaries, such that when two polygons
do not overlap or two lines do not cross, no point exist that belong to
more than one of the geometries. (This latter form, half-closed, is
not present n the OpenGIS "simple feature access" (SFA) standard, or DE9-IM on
which that is based). A value of -1 does not set the model, leaving the
S2 default (HALF-CLOSED). The default values for \code{s2_contains} (0)
and covers/coveredby (2) correspond to the SFA standard specification
of these operators.
}
\seealso{
BigQuery's geography function reference:
\itemize{
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_contains}{ST_CONTAINS}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_coveredby}{ST_COVEREDBY}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_covers}{ST_COVERS}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_disjoint}{ST_DISJOINT}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_equals}{ST_EQUALS}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_intersects}{ST_INTERSECTS}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_intersectsbox}{ST_INTERSECTSBOX}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_touches}{ST_TOUCHES}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_within}{ST_WITHIN}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_dwithin}{ST_DWITHIN}
}
}
