% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s2_data_example_wkt}
\alias{s2_data_example_wkt}
\title{Example Geometries}
\format{
An object of class \code{list} of length 29.
}
\usage{
s2_data_example_wkt
}
\description{
These geometries are toy examples useful for testing various coordinate
shuffling operations in the s2 package.
}
\keyword{datasets}
