% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s2_data_world_borders}
\alias{s2_data_world_borders}
\alias{s2_data_timezones}
\alias{s2data_countries}
\alias{s2data_timezones}
\title{Low-resolution world boundaries and timezones}
\format{
A data.frame with columns \code{name} (character), and
\code{geometry} (wk_wkb)

An object of class \code{data.frame} with 120 rows and 2 columns.
}
\source{
\code{\link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries()}}
}
\usage{
s2_data_world_borders

s2_data_timezones

s2data_countries(name = NULL)

s2data_timezones(utc_offset_min = NULL, utc_offset_max = utc_offset_min)
}
\arguments{
\item{name}{The name of of a country or continent, or \code{NULL}
for all countries.}
}
\description{
Well-known binary versions of the \href{https://www.naturalearthdata.com/}{Natural Earth}
low-resolution world boundaries and timezone boundaries.
}
\examples{
head(s2data_countries())
s2data_countries("Germany")
s2data_countries("Europe")

}
\keyword{datasets}
