% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2-earth.R
\name{s2_earth_radius_meters}
\alias{s2_earth_radius_meters}
\title{Earth constants}
\usage{
s2_earth_radius_meters()
}
\description{
According to Yoder (1995), the radius of the earth is
6371.01 km.
}
\examples{
s2_earth_radius_meters()

}
\references{
Yoder, C.F. 1995. "Astrometric and Geodetic Properties of Earth and the
Solar System" in Global Earth Physics, A Handbook of Physical Constants,
AGU Reference Shelf 1, American Geophysical Union, Table 2.
\url{https://doi.org/10.1029/RF001p0001}
}
