% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-constructors-formatters.R
\name{s2_geogpoint}
\alias{s2_geogpoint}
\alias{s2_makeline}
\alias{s2_makepolygon}
\alias{s2_makepolygonoriented}
\alias{s2_geogfromtext}
\alias{s2_geogfromwkb}
\alias{s2_astext}
\alias{s2_asbinary}
\title{Create and format geography vectors}
\usage{
s2_geogpoint(longitude, latitude)

s2_makeline(longitude, latitude)

s2_makepolygon(longitude, latitude)

s2_makepolygonoriented(longitude, latitude)

s2_geogfromtext(wkt_string, oriented = FALSE)

s2_geogfromwkb(wkb_bytes, oriented = FALSE)

s2_astext(x)

s2_asbinary(x, endian = match(.Platform$endian, c("big", "little")) - 1)
}
\arguments{
\item{longitude, latitude}{Vectors of latitude and longitude}

\item{wkt_string}{Well-known text}

\item{oriented}{TRUE if polygon ring directions are known to be correct
(i.e., exterior rings are defined counter clockwise and interior
rings are defined clockwise).}

\item{wkb_bytes}{A \code{list()} of \code{raw()}}

\item{x}{An object that can be interpreted as an S2 type.}

\item{endian}{0 for big endian and 1 for little endian.}
}
\description{
Create and format geography vectors
}
\seealso{
BigQuery's geography function reference:
\itemize{
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_geogpoint}{ST_GEOGPOINT}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_makeline}{ST_MAKELINE}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_makepolygon}{ST_MAKEPOLYGON}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_makepolygonoriented}{ST_MAKEPOLYGONORIENTED}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_geogfromtext}{ST_GEOGFROMTEXT}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_geogfromwkb}{ST_GEOGFROMWKB}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_astext}{ST_ASTEXT}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_asbinary}{ST_ASBINARY}
}
}
