% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2-geography.R
\name{s2_geography}
\alias{s2_geography}
\alias{s2_geography.s2_geography}
\alias{s2_geography.wk_wkb}
\alias{s2_geography.logical}
\alias{s2_geography.character}
\title{Create an S2 Geography Vector}
\usage{
s2_geography(x, ...)

\method{s2_geography}{s2_geography}(x, ...)

\method{s2_geography}{wk_wkb}(x, ..., oriented = FALSE)

\method{s2_geography}{logical}(x, ...)

\method{s2_geography}{character}(x, ..., oriented = FALSE)
}
\arguments{
\item{x}{An object that can be converted to an s2_geography vector}

\item{...}{Unused}

\item{oriented}{TRUE if polygon ring directions are known to be correct
(i.e., exterior rings are defined counter clockwise and interior
rings are defined clockwise).}
}
\value{
A \code{\link[=new_s2_xptr]{new_s2_xptr()}} with class s2_geography
}
\description{
Create an S2 Geography Vector
}
