% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-accessors.R
\name{s2_iscollection}
\alias{s2_iscollection}
\alias{s2_dimension}
\alias{s2_numpoints}
\alias{s2_isempty}
\alias{s2_area}
\alias{s2_length}
\alias{s2_perimeter}
\alias{s2_x}
\alias{s2_y}
\alias{s2_distance}
\alias{s2_maxdistance}
\title{S2 Geography Accessors}
\usage{
s2_iscollection(x)

s2_dimension(x)

s2_numpoints(x)

s2_isempty(x)

s2_area(x)

s2_length(x)

s2_perimeter(x)

s2_x(x)

s2_y(x)

s2_distance(x, y, radius = s2earth_radius_meters())

s2_maxdistance(x, y, radius = s2earth_radius_meters())
}
\arguments{
\item{x, y}{An object that can be interpreted as an S2 type.}

\item{radius}{Radius of the earth. Defaults to the average radius of
the earth as defined by \code{\link[=s2earth_radius_meters]{s2earth_radius_meters()}}.}
}
\description{
S2 Geography Accessors
}
\seealso{
BigQuery's geography function reference:
\itemize{
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_iscollection}{ST_ISCOLLECTION}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_dimension}{ST_DIMENSION}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_numpoints}{ST_NUMPOINTS}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_area}{ST_AREA}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_length}{ST_LENGTH}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_perimeter}{ST_PERIMETER}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_x}{ST_X}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_y}{ST_Y}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_distance}{ST_DISTANCE}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_maxdistance}{ST_MAXDISTANCE}
}
}
