% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s2_nc_wkb}
\alias{s2_nc_wkb}
\title{Geometry of the nc dataset of package sf}
\format{
A list with components \code{CNTY_ID} (integer), \code{NAME} (character), and
\code{geometry} (wk_wkb)
}
\source{
\url{http://r-spatial.github.io/sf/}
}
\usage{
s2_nc_wkb
}
\description{
A subset of \link[sf:sf]{sf's} North Carolina dataset in WKB format.
}
\examples{
nc <- s2polygon(s2_nc_wkb$geometry)
head(nc)

}
\keyword{datasets}
