% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2-state.R
\name{s2_snap_default}
\alias{s2_snap_default}
\alias{s2_model_default}
\title{Default states}
\usage{
s2_snap_default()

s2_model_default()
}
\description{
Default states
}
\examples{
s2_snap_default()
s2_model_default()

}
