% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s2_wrld_simpl_wkb}
\alias{s2_wrld_simpl_wkb}
\alias{s2data_country}
\title{Low-resolution world boundaries}
\format{
A list with components \code{ISO3} (character), \code{NAME} (character), and
\code{geometry} (wk_wkb)
}
\source{
The \link[mapview:mapview-package]{mapview package}.
}
\usage{
s2_wrld_simpl_wkb

s2data_country(name = NULL)
}
\arguments{
\item{name}{The name or ISO3 identifier of a country, or \code{NULL}
for all countries.}
}
\description{
A modified version of \link[maptools:wrld_simpl]{maptools'} \code{wrld_simpl} dataset
in WKB format. Use the \code{\link[=s2data_country]{s2data_country()}} helper to load an \code{\link[=s2polygon]{s2polygon()}} vector for
a given country.
}
\examples{
#world <- s2polygon(s2_wrld_simpl_wkb$geometry)
#head(world)

#s2data_country("Germany")

}
\keyword{datasets}
