% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s2BuildPolygonsLayer}
\alias{s2BuildPolygonsLayer}
\title{Build s2Polygons Layers}
\usage{
s2BuildPolygonsLayer(ptrs, b_ptr)
}
\arguments{
\item{ptrs}{list with S2Polygons pointers}

\item{b_ptr}{external pointer with existing build, or NULL}
}
\description{
Build s2Polygons Layers
}
