% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s2cap}
\alias{s2cap}
\title{Construct a S2Cap using axis and height from R}
\usage{
s2cap(axis, height)
}
\arguments{
\item{axis}{A numeric vector with three entries represtenting the direction axis.}

\item{height}{Single numeric representing the height of the cap.}
}
\description{
Constructs a S2Cap
}
