% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s2cap_contains_point}
\alias{s2cap_contains_point}
\title{Point in s2cap test}
\usage{
s2cap_contains_point(points, cap)
}
\arguments{
\item{points}{A three column matrix represtenting the points.}

\item{cap}{Named list containing axis and height of cap.}
}
\description{
Test whether points on the sphere are contained in a cap
}
