% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S2Cap_GetRectBound}
\alias{S2Cap_GetRectBound}
\title{Bounding latitude and longitude rectangle for spherical cap}
\usage{
S2Cap_GetRectBound(x)
}
\arguments{
\item{x}{cap}
}
\description{
Bounding latitude and longitude rectangle for spherical cap
}
