% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S2CapFromAxisHeight}
\alias{S2CapFromAxisHeight}
\title{Construct a S2Cap using axis and height}
\usage{
S2CapFromAxisHeight(axis, height)
}
\arguments{
\item{axis}{A numeric vector with three entries represtenting the direction axis.}

\item{height}{Single numeric representing the height of the cap.}
}
\description{
Constructs a S2Cap from axis and height
}
