% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2regioncoverer.R
\name{S2Cell}
\alias{S2Cell}
\title{Make a list of S2Cells}
\usage{
S2Cell(x)
}
\arguments{
\item{x}{Input to create cells from. Currently only an object of class
S2CellId is supported.}
}
\description{
Make a list of S2Cells
}
