% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S2CellId_ToPoint}
\alias{S2CellId_ToPoint}
\title{Convert S2CellId to a S2Point}
\usage{
S2CellId_ToPoint(x)
}
\arguments{
\item{x}{Object of class S2CellId.}
}
\value{
Three-column matrix reprensenting the points..
}
\description{
Convert S2CellId to a S2Point
}
