% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S2CellId_ToString}
\alias{S2CellId_ToString}
\title{Make a vector of S2CellId strings}
\usage{
S2CellId_ToString(x)
}
\arguments{
\item{x}{A charecter vector with S2CellIds (in token form).}
}
\value{
A character vector with S2CellId strings.
}
\description{
Make a vector of S2CellId strings
}
