% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S2CellIdFromLatLng}
\alias{S2CellIdFromLatLng}
\title{Make a Vector of S2CellIds From S2LatLng on the Sphere}
\usage{
S2CellIdFromLatLng(x, level = 30L)
}
\arguments{
\item{x}{List of S2LatLng}

\item{level}{Integer between 0 and 30 (incl).}
}
\value{
An object of class `S2CellId`.
}
\description{
Create a vector of S2CellIds corresponding to the cells at the given level 
containing the given coordinates. The default level (30) corresponds to leaf
cells (finest level).
}
