% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S2CellIdFromPoint}
\alias{S2CellIdFromPoint}
\title{Make a Vector of S2CellIds From Points on the Sphere}
\usage{
S2CellIdFromPoint(x, level = 30L)
}
\arguments{
\item{x}{Three-column matrix reprensenting the points.}

\item{level}{Integer between 0 and 30 (incl).}
}
\value{
An object of class `S2CellId`.
}
\description{
Create a vector of S2CellIds corresponding to the cells at the given level 
containing the given points. The default level (30) corresponds to leaf
cells (finest level).
}
