% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s2Intersects}
\alias{s2Intersects}
\title{Return indices of intersecting S2Polygons}
\usage{
s2Intersects(x, y)
}
\arguments{
\item{x}{list with S2Polygons pointers}

\item{y}{list with S2Polygons pointers}
}
\description{
Return indices of intersecting S2Polygons
}
