% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2latlng.R
\name{s2latlng}
\alias{s2latlng}
\alias{s2latlng.s2latlng}
\alias{s2latlng.s2point}
\alias{s2latlng.numeric}
\alias{s2latlng.matrix}
\alias{as.data.frame.s2latlng}
\alias{as.matrix.s2latlng}
\title{Create an s2latlng vector}
\usage{
s2latlng(lat, ...)

\method{s2latlng}{s2latlng}(lat, ...)

\method{s2latlng}{s2point}(lat, ...)

\method{s2latlng}{numeric}(lat, lng, ...)

\method{s2latlng}{matrix}(lat, ...)

\method{as.data.frame}{s2latlng}(x, ...)

\method{as.matrix}{s2latlng}(x, ...)
}
\arguments{
\item{lat, lng}{Vectors of latitude and longitude values in degrees.}

\item{...}{Unused}

\item{x}{A \code{\link[=s2latlng]{s2latlng()}} vector}
}
\value{
A \code{\link[=new_s2xptr]{new_s2xptr()}} with class s2latlng
}
\description{
This class represents a latitude and longitude on the Earth's surface.
Most calculations in S2 convert this to a \code{\link[=s2point]{s2point()}}, which is a
unit vector representation of this value.
}
\examples{
s2latlng(45, -64) # Halifax, Nova Scotia!
as.data.frame(s2latlng(45, -64))

}
