% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-compat.R
\name{s2latlng.sfc}
\alias{s2latlng.sfc}
\alias{s2polyline.sfc}
\alias{s2polygon.sfc}
\title{S2-sf compatibility functions}
\usage{
\method{s2latlng}{sfc}(lng, ...)

\method{s2polyline}{sfc}(lng, ...)

\method{s2polygon}{sfc}(lng, ...)
}
\arguments{
\item{lng, x}{Objects}

\item{...}{Unused}
}
\description{
S2-sf compatibility functions
}
