% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S2LatLngRect}
\alias{S2LatLngRect}
\title{Create a rectangle of latitude and longitude on the sphere}
\usage{
S2LatLngRect(lo, hi)
}
\arguments{
\item{lo}{Latitude and longitude (in that order) in degrees of lower left corner.}

\item{hi}{Latitude and longitude (in that order) in degrees of upper right corner.}
}
\description{
Create a rectangle of latitude and longitude on the sphere
}
