% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S2LatLngRect_area}
\alias{S2LatLngRect_area}
\title{Area of s2latlngrect}
\usage{
S2LatLngRect_area(x)
}
\arguments{
\item{x}{Named list containing axis and height of cap.}
}
\description{
Area of a S2LatLngRect.
}
