% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s2makepoint}
\alias{s2makepoint}
\alias{s2MakePoint}
\title{convert R vector with coordinates (lon,lat) into S2Point ptr}
\usage{
s2MakePoint(pt)
}
\arguments{
\item{pt}{numeric; length 2, longitude latitude}
}
\description{
convert R vector with coordinates (lon,lat) into S2Point ptr
}
