% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s2makepolygons}
\alias{s2makepolygons}
\alias{s2IsValid}
\alias{s2GetArea}
\alias{s2MakePolygon}
\alias{s2GetPolygon}
\title{convert R list of coordinate matrices (lon,lat) into S2Polygon ptr}
\usage{
s2IsValid(ptrs)

s2GetArea(ptrs)

s2MakePolygon(mat, oriented = FALSE)

s2GetPolygon(ptrs)
}
\arguments{
\item{ptrs}{R list with external references (pointers) to S2Polygon objects}

\item{mat}{two-column matrix with longitude in first, latitude in second column}

\item{oriented}{logical; if TRUE, rings are guaranteed to be oriented (e.g. read
by \code{read_sf} using \code{check_ring_dir=TRUE}), meaning CCW exterior rings
and CW holes; if FALSE, rings are normalized and holes are deduced from degree of nesting.}
}
\description{
convert R list of coordinate matrices (lon,lat) into S2Polygon ptr
}
