% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s2makepolyline}
\alias{s2makepolyline}
\alias{s2MakePolyline}
\alias{s2GetPolyline}
\title{convert R matrix with coordinates (lon,lat) into S2Polyline ptr}
\usage{
s2MakePolyline(pts)

s2GetPolyline(ptrs)
}
\arguments{
\item{pts}{2-column numeric matrix with lng,lat line vertices}

\item{ptrs}{R list with external references (pointers) to S2Polyline objects}
}
\description{
convert R matrix with coordinates (lon,lat) into S2Polyline ptr
}
