% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s2measures}
\alias{s2measures}
\alias{s2polyline_lengths}
\alias{s2polygon_areas}
\title{Get lengths of s2polylines}
\usage{
s2polyline_lengths(s2polyline, degrees = FALSE)

s2polygon_areas(s2polygon)
}
\arguments{
\item{s2polyline}{object of class \link{s2polyline}}

\item{degrees}{logical; if \code{FALSE}, return length in radians, else in degrees}

\item{s2polygon}{object of class \link{s2polygon}}
}
\description{
Get lengths of s2polylines

Get area of s2polygons
}
\details{
note that all s2 measures are on the unit sphere, and need rescaling
by the Earth's radius to get measures on the Earth
}
