% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s2ops}
\alias{s2ops}
\alias{s2Intersects}
\alias{s2IsValid}
\alias{s2GetArea}
\title{Geometry operators for s2 geometries}
\usage{
s2Intersects(x, y, polygons = TRUE)

s2IsValid(ptrs, polygons = TRUE)

s2GetArea(ptrs)
}
\arguments{
\item{x}{list with S2Polygons or S2Polyline pointers}

\item{y}{list with S2Polygons or S2Polyline pointers}

\item{polygons}{logical; if TRUE, x and y are S2Polygon, otherwise S2Polyline}

\item{ptrs}{list of S2Polygon or S2Polyline pointers}
}
\description{
Geometry operators for s2 geometries
}
