% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2point.R
\name{s2point}
\alias{s2point}
\alias{s2point.s2point}
\alias{s2point.numeric}
\alias{s2point.matrix}
\alias{as.data.frame.s2point}
\alias{as.matrix.s2point}
\title{Create an s2point vector}
\usage{
s2point(x, ...)

\method{s2point}{s2point}(x, ...)

\method{s2point}{numeric}(x, y, z, ...)

\method{s2point}{matrix}(x, ...)

\method{as.data.frame}{s2point}(x, ...)

\method{as.matrix}{s2point}(x, ...)
}
\arguments{
\item{...}{Unused}

\item{lat, lng}{Vectors of latitude and longitude values in degrees.}
}
\value{
A \code{\link[=new_s2xptr]{new_s2xptr()}} with class s2point
}
\description{
Create an s2point vector
}
