% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S2PointVector_interpolate}
\alias{S2PointVector_interpolate}
\title{Interpolation of points on unit sphere}
\usage{
S2PointVector_interpolate(x, eps)
}
\arguments{
\item{x}{Matrix with three columns representing the points.}

\item{eps}{Strictly positive real number. Values greater than or equal to pi
correspond to no interpolation.}
}
\description{
Given a sequence of points on the unit sphere add interpolating points so
consecutive points are within distance `eps` of each other.
}
