% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2polygon.R
\name{s2polygon}
\alias{s2polygon}
\alias{s2polygon.s2polygon}
\alias{s2polygon.s2polyline}
\title{Create an s2polygon vector}
\usage{
s2polygon(x, ...)

\method{s2polygon}{s2polygon}(x, ...)

\method{s2polygon}{s2polyline}(x, ..., oriented = FALSE, check = TRUE)
}
\arguments{
\item{x}{An object from which an s2polygon can be created}

\item{...}{Unused}
}
\value{
A \code{\link[=new_s2xptr]{new_s2xptr()}} with class s2polygon
}
\description{
Create an s2polygon vector
}
