% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2polygon.R
\name{S2Polygon_Contains}
\alias{S2Polygon_Contains}
\title{Test of Containment in S2Polygon}
\usage{
S2Polygon_Contains(poly, x, approx = TRUE)
}
\arguments{
\item{poly}{Named list containing an entry called `loops` containing a list
of polygon loops represented by three-column matrices.}

\item{x}{Object(s) to test for containment in `poly`. At the moment only
points are handled and they must be specified as a three-column matrix.}

\item{approx}{Logical to use approximate testing of point in polygon (allows
points very slightly outside the polygon). Useful for allowing points
directly on the border.}
}
\description{
Test whether the given object(s) on the sphere are contained in a polygon. At
the moment only point containment can be tested.
}
