% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s2polygon_contains_point}
\alias{s2polygon_contains_point}
\title{Point in s2polygon test}
\usage{
s2polygon_contains_point(points, poly, approx = TRUE)
}
\arguments{
\item{points}{A three column matrix represtenting the points.}

\item{poly}{List of polygon loops represented by three column matrices.}

\item{approx}{Logical to use approximate testing of point in polygon (allows
points very slightly outside the polygon). Useful for allowing points
directly on the border.}
}
\description{
Test whether points on the sphere are contained in a polygon on the sphere
}
