% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s2polygon_union}
\alias{s2polygon_union}
\alias{s2polygon_intersection}
\title{Union and intersection of two s2polygons}
\usage{
s2polygon_union(x, y)
}
\arguments{
\item{x}{List of loops represented by three column matrices.}

\item{y}{List of loops represented by three column matrices.}
}
\description{
Given two lists of three column matrices representing S2Polygons calculate
the union or intersection. Assumes the polygons have already been validated
and put into the correct format by [`s2polygon`]. Note the input is only the
loops/rings of two polygons not the entire objects with areas and hole
indicators.
}
