% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S2Polygons_area}
\alias{S2Polygons_area}
\title{compute areas for a list of s2polygons}
\usage{
S2Polygons_area(x)
}
\arguments{
\item{x}{List of list of loops represented by three-column matrices.}
}
\description{
this function is equivalent to \link[sf]{st_area}, in that it
returns a numeric vector with polygon areas
}
\details{
the area is on the unit sphere, in [0, 4 * pi]
}
