% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S2Polygons_centroid}
\alias{S2Polygons_centroid}
\title{compute centroids for a list of s2polygons}
\usage{
S2Polygons_centroid(x)
}
\arguments{
\item{x}{List of list of loops represented by three-column matrices.}
}
\description{
this function is equivalent to \link[sf]{st_centroid}, in that it
returns a numeric vector with polygon centroids
}
