% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S2Polygons_intersect}
\alias{S2Polygons_intersect}
\title{for two sets of s2polygons, which ones intersect?}
\usage{
S2Polygons_intersect(x, y)
}
\arguments{
\item{x}{List of list of loops represented by three-column matrices.}

\item{y}{List of list of loops represented by three-column matrices.}
}
\description{
this function is equivalent to \link[sf]{st_intersects}, in that it
returns a sparse matrix with indexes for pairs of intersecting polygons
}
