% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S2Polygons_intersection}
\alias{S2Polygons_intersection}
\title{intersection of sets of s2polygons}
\usage{
S2Polygons_intersection(x, y)
}
\arguments{
\item{x}{List of list of loops represented by three-column matrices.}

\item{y}{List of list of loops represented by three-column matrices.}
}
\description{
this function generalizes S2Polygon_intersection, by allowing two lists of
polygons, each list element equal to a polygon of that function.
}
