% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S2Polyline_dist}
\alias{S2Polyline_dist}
\title{Distance from points to line on sphere}
\usage{
S2Polyline_dist(line, x)
}
\arguments{
\item{line}{Line represented by a sequence of vertices given as a single
three-column matrices with one line per vertex.}

\item{x}{Points represented by three-column matrices.}
}
\description{
Distance from points to line on sphere.
}
