% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{snaplevel}
\alias{snaplevel}
\alias{s2_set_snaplevel}
\alias{s2_get_snaplevel}
\title{Set snap level for polygons and polylines}
\usage{
s2_set_snaplevel(snap = -1L)

s2_get_snaplevel(snap = -1L)
}
\arguments{
\item{snap}{integer, snap level (max. 30, negative prevents snapping)}
}
\value{
s2_set_snaplevel returns the old value of snap level
}
\description{
Set snap level for polygons and polylines
}
