\name{interval_test}
\alias{interval_test}
\title{Interval test}
\description{Test whether a number is inside a given interval (dummy function to test package).}
\usage{
interval_test(x, a, b)
}
\arguments{
\item{x}{Numeric. Point to be tested.}

\item{a}{Numeric. Lower limit of interval.}

\item{b}{Numeric. Upper limit of interval.}
}
\value{
Logical.
}
