% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2xptr.R
\name{new_s2xptr}
\alias{new_s2xptr}
\alias{validate_s2xptr}
\alias{rep.s2xptr}
\alias{rep_len.s2xptr}
\title{Create vectors of XPtr objects}
\usage{
new_s2xptr(x = list(), class = character())

validate_s2xptr(x)

\method{rep}{s2xptr}(x, ...)

rep_len.s2xptr(x, ...)
}
\arguments{
\item{x}{A bare \code{list()} of external pointers}

\item{class}{A character vector subclass}

\item{...}{Unused}
}
\value{
An object of class s2xptr
}
\description{
Create vectors of XPtr objects
}
