% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{S2-classes}
\alias{S2-classes}
\alias{S1Angle}
\alias{S2Cap}
\alias{S2LatLng}
\alias{S2LatLngRect}
\alias{S2Loop}
\alias{S2Point}
\alias{S2Polygon}
\alias{S2PolygonBuilder}
\alias{S2PolygonBuilderOptions}
\title{Wrapper classes for corresponding classes in Google's S2 geometry library}
\description{
These classes are reference classes that wrap C++ classes of the
  same name.
}
