% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2-package.R
\docType{package}
\name{s2-package}
\alias{s2}
\alias{s2-package}
\title{s2: Spherical Geometry Operators Using the S2 Geometry Library}
\description{
Provides R bindings for Google's s2 library for geometric calculations on the sphere. High-performance constructors and exporters provide high compatibility with existing spatial packages, transformers construct new geometries from existing geometries, predicates provide a means to select geometries based on spatial relationships, and accessors extract information about geometries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-spatial.github.io/s2/}
  \item \url{https://github.com/r-spatial/s2}
  \item \url{http://s2geometry.io/}
  \item Report bugs at \url{https://github.com/r-spatial/s2/issues}
}

}
\author{
\strong{Maintainer}: Edzer Pebesma \email{edzer.pebesma@uni-muenster.de} (\href{https://orcid.org/0000-0001-8049-7069}{ORCID})

Authors:
\itemize{
  \item Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})
  \item Ege Rubak \email{rubak@math.aau.dk}
}

Other contributors:
\itemize{
  \item Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu} (configure script) [contributor]
  \item  Google, Inc. (Original s2geometry.io source code) [copyright holder]
}

}
\keyword{internal}
