% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2.R
\docType{package}
\name{s2-package}
\alias{s2}
\alias{s2-package}
\title{Google's s2 library for geometry on the sphere}
\description{
R bindings for Google's s2 library for geometric calculations on
the sphere.
}
\details{
This package aims at providing bindings for the C++ library s2 for geometry
on the sphere. The C++ library was originally developed by Google under the
Apache license. At the moment very few things are exposed and the main
concern is to make the C++ code compile with the R toolchain on the three
major platforms and have the package accepted on CRAN. Once this proof of
concept has been achieved more of the underlying features will be made
available to the useR.
}

