% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{s2_boundary}
\alias{s2_boundary}
\alias{s2_centroid}
\alias{s2_snaptogrid}
\alias{s2_union}
\alias{s2_closestpoint}
\alias{s2_difference}
\alias{s2_intersection}
\alias{s2_union_agg}
\alias{s2_centroid_agg}
\title{S2 Geography Transformations}
\usage{
s2_boundary(x)

s2_centroid(x)

s2_snaptogrid(x)

s2_union(x, y = NULL)

s2_closestpoint(x, y)

s2_difference(x, y)

s2_intersection(x, y)

s2_union_agg(x)

s2_centroid_agg(x)
}
\arguments{
\item{x}{An object that can be interpreted as an S2 type.}

\item{y}{An object that can be interpreted as an S2 type.}
}
\description{
S2 Geography Transformations
}
\seealso{
BigQuery's geography function reference:
\itemize{
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_boundary}{ST_BOUNDARY}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_centroid}{ST_CENTROID}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_closestpoint}{ST_CLOSESTPOINT}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_difference}{ST_DIFFERENCE}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_intersection}{ST_INTERSECTION}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_snaptogrid}{ST_SNAPTOGRID}
\item \href{https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_union}{ST_UNION}
}
}
