% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2-cell-union.R
\name{s2_cell_union}
\alias{s2_cell_union}
\alias{as_s2_geography.s2_cell_union}
\alias{as_s2_cell_union}
\alias{as_s2_cell_union.s2_cell_union}
\alias{as_s2_cell_union.s2_cell}
\alias{as_s2_cell_union.character}
\title{Create S2 Cell Union vectors}
\usage{
s2_cell_union(x = list())

\method{as_s2_geography}{s2_cell_union}(x, ...)

as_s2_cell_union(x, ...)

\method{as_s2_cell_union}{s2_cell_union}(x, ...)

\method{as_s2_cell_union}{s2_cell}(x, ...)

\method{as_s2_cell_union}{character}(x, ...)
}
\arguments{
\item{x}{A \code{list()} of \code{\link[=s2_cell]{s2_cell()}} vectors.}

\item{...}{Passed to S3 methods}
}
\value{
An object of class "s2_cell_union".
}
\description{
Create S2 Cell Union vectors
}
