% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s2_data_world_borders}
\alias{s2_data_world_borders}
\alias{s2data_countries}
\title{Low-resolution world boundaries}
\format{
A data.frame with columns \code{name} (character), and
\code{geometry} (wk_wkb)
}
\source{
\code{\link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries()}}
}
\usage{
s2_data_world_borders

s2data_countries(name = NULL)
}
\arguments{
\item{name}{The name or of a country, or \code{NULL}
for all countries.}
}
\description{
A well-known binary version of the \href{https://www.naturalearthdata.com/}{Natural Earth}
low-resolution world boundaries.
}
\examples{
head(s2data_countries())
s2data_countries("Germany")
s2data_countries("Canada")

}
\keyword{datasets}
