% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2-point.R
\name{s2_point}
\alias{s2_point}
\alias{as_s2_point}
\alias{as_s2_point.s2_point}
\alias{as_s2_point.s2_latlng}
\alias{as_s2_point.matrix}
\alias{as.data.frame.s2_point}
\alias{as.matrix.s2_point}
\title{Create an s2_point vector}
\usage{
s2_point(x, y, z)

as_s2_point(x, ...)

\method{as_s2_point}{s2_point}(x, ...)

\method{as_s2_point}{s2_latlng}(x, ...)

\method{as_s2_point}{matrix}(x, ...)

\method{as.data.frame}{s2_point}(x, ...)

\method{as.matrix}{s2_point}(x, ...)
}
\arguments{
\item{x, y, z}{Vectors of latitude and longitude values in degrees.}

\item{...}{Unused}
}
\value{
An object with class s2_point
}
\description{
In S2 terminology, a "point" is a 3-dimensional unit vector, as opposed
to a \code{\link[=as_s2_latlng]{as_s2_latlng()}}, which is a position on the Earth's surface.
}
\examples{
latlng <- as_s2_latlng(45, -64) # Halifax, Nova Scotia!
as_s2_point(latlng)
as.data.frame(as_s2_point(latlng))

}
