% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2cap.R
\name{S2Cap_Contains}
\alias{S2Cap_Contains}
\title{Test of Containment in S2Cap}
\usage{
S2Cap_Contains(cap, x)
}
\arguments{
\item{cap}{Named list containing axis and height of cap.}

\item{x}{Object(s) to test for containment in `cap`. At the moment only
points are handled and they must be specified as a three-column matrix.}
}
\description{
Test whether the given object(s) on the sphere are contained in a cap. At the
moment only point containment can be tested.
}
\examples{
cap <- list(axis = c(0,1,0), height = 0.1)
S2Cap_Contains(cap, matrix(c(0,1,0),ncol=3))
S2Cap_Contains(cap, matrix(c(1,0,0),ncol=3))
}
