% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2geography.R
\name{s2geography}
\alias{s2geography}
\alias{s2geography.s2geography}
\alias{s2geography.wk_wkb}
\alias{s2geography.logical}
\alias{s2geography.character}
\title{Create an S2 Geography Vector}
\usage{
s2geography(x, ...)

\method{s2geography}{s2geography}(x, ...)

\method{s2geography}{wk_wkb}(x, ..., oriented = FALSE)

\method{s2geography}{logical}(x, ...)

\method{s2geography}{character}(x, ..., oriented = FALSE)
}
\arguments{
\item{x}{An object that can be converted to an s2geography vector}

\item{...}{Unused}

\item{oriented}{TRUE if polygon ring directions are known to be correct
(i.e., exterior rings are defined counter clockwise and interior
rings are defined clockwise).}
}
\value{
A \code{\link[=new_s2xptr]{new_s2xptr()}} with class s2geography
}
\description{
Create an S2 Geography Vector
}
