% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2latlng.R
\name{s2latlng}
\alias{s2latlng}
\alias{s2latlng.s2latlng}
\alias{s2latlng.numeric}
\alias{s2latlng.matrix}
\alias{as.data.frame.s2latlng}
\title{Create an s2latlng vector}
\usage{
s2latlng(lat, ...)

\method{s2latlng}{s2latlng}(lat, ...)

\method{s2latlng}{numeric}(lat, lng, ...)

\method{s2latlng}{matrix}(lat, ...)

\method{as.data.frame}{s2latlng}(x, ...)
}
\arguments{
\item{lat, lng}{Vectors of latitude and longitude values in degrees.}

\item{...}{Unused}
}
\value{
A \code{\link[=new_s2xptr]{new_s2xptr()}} with class s2latlng
}
\description{
Create an s2latlng vector
}
