% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-compat.R
\name{s2latlng.sfc}
\alias{s2latlng.sfc}
\alias{s2polyline.sfc}
\alias{s2polygon.sfc}
\title{S2-sf compatibility functions}
\usage{
\method{s2latlng}{sfc}(lat, ...)

\method{s2polyline}{sfc}(x, ...)

\method{s2polygon}{sfc}(x, ...)
}
\arguments{
\item{...}{Unused}

\item{x}{A \code{\link[=s2latlng]{s2latlng()}} vector}
}
\description{
S2-sf compatibility functions
}
