% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s2ops}
\alias{s2ops}
\alias{s2Intersects}
\alias{s2IsValid}
\alias{s2GetArea}
\title{Geometry operators for s2 geometries}
\usage{
s2Intersects(x, y, polygons = TRUE)

s2IsValid(ptrs, polygon = TRUE)

s2GetArea(ptrs)
}
\arguments{
\item{x}{list with S2Polygons pointers}

\item{y}{list with S2Polygons pointers}
}
\description{
Geometry operators for s2 geometries
}
