% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s2ops}
\alias{s2ops}
\alias{s2polygon_intersects}
\alias{s2polyline_intersects}
\alias{s2polygon_is_valid}
\alias{s2polyline_is_valid}
\title{Geometry operators for s2 geometries}
\usage{
s2polygon_intersects(x, y)

s2polyline_intersects(x, y)

s2polygon_is_valid(ptrs)

s2polyline_is_valid(ptrs)
}
\arguments{
\item{x}{list with S2Polygons or S2Polyline pointers}

\item{y}{list with S2Polygons or S2Polyline pointers}

\item{ptrs}{list of S2Polygon or S2Polyline pointers}
}
\description{
Geometry operators for s2 geometries
}
