% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2-point.R
\name{s2point}
\alias{s2point}
\alias{s2point.s2point}
\alias{s2point.s2latlng}
\alias{s2point.numeric}
\alias{s2point.matrix}
\alias{as.data.frame.s2point}
\alias{as.matrix.s2point}
\title{Create an s2point vector}
\usage{
s2point(x, ...)

\method{s2point}{s2point}(x, ...)

\method{s2point}{s2latlng}(x, ...)

\method{s2point}{numeric}(x, y, z, ...)

\method{s2point}{matrix}(x, ...)

\method{as.data.frame}{s2point}(x, ...)

\method{as.matrix}{s2point}(x, ...)
}
\arguments{
\item{x, y, z}{Vectors of latitude and longitude values in degrees.}

\item{...}{Unused}
}
\value{
A \code{\link[=new_s2xptr]{new_s2xptr()}} with class s2point
}
\description{
In S2 terminology, a "point" is a 3-dimensional unit vector, as opposed
to a \code{\link[=s2latlng]{s2latlng()}}, which is a position on the Earth's surface.
}
\examples{
latlng <- s2latlng(45, -64) # Halifax, Nova Scotia!
s2point(latlng)
as.data.frame(s2point(latlng))

}
