% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s2polygon}
\alias{s2polygon}
\title{Create an object of class s2polygon}
\usage{
s2polygon(x, validate = TRUE)
}
\arguments{
\item{x}{List of loops}

\item{use_loops}{logical to add ed}
}
\description{
At the moment this function is very rough around the edges...
}
