% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2polyline.R
\name{s2polyline}
\alias{s2polyline}
\alias{s2polyline.s2polyline}
\alias{s2polyline.s2latlng}
\alias{s2polyline.matrix}
\alias{s2latlng.s2polyline}
\alias{as.matrix.s2polyline}
\alias{as.data.frame.s2polyline}
\title{Create an s2polyline vector}
\usage{
s2polyline(x, ...)

\method{s2polyline}{s2polyline}(x, ...)

\method{s2polyline}{s2latlng}(x, ...)

\method{s2polyline}{matrix}(x, ...)

\method{s2latlng}{s2polyline}(lat, ...)

\method{as.matrix}{s2polyline}(x, ...)

\method{as.data.frame}{s2polyline}(x, ...)
}
\arguments{
\item{x}{An object from which an s2polyline can be created}

\item{...}{Unused}

\item{lat}{Vectors of latitude and longitude values in degrees.}
}
\value{
A \code{\link[=new_s2xptr]{new_s2xptr()}} with class s2polyline
}
\description{
S2 polylines represent a collection of vertices composed of
\code{\link[=s2latlng]{s2latlng()}} objects.
}
\examples{
# construct and export a single polyline using s2latlng()
(polyline <- s2polyline(s2latlng(45:50, -64)))
s2latlng(polyline)

}
